move dummy generation of tables into tsql lib
