+++ lib/Cargo.toml
@@ -12,9 +12,14 @@ repository.workspace = true
 anyhow = "1.0.75"
 nom = "7.1.3"
 static_assertions = "1.1.0"
+hmac-sha256 = { version = "1.1.7", optional = true }
 
 [dev-dependencies]
 criterion = { version = "0.4", features = ["html_reports"] }
+
+[features]
+default = ["generate"]
+generate = ["dep:hmac-sha256"]
 
 [[bench]]
 name = "parsing_tables"
