+++ lib/src/lib.rs
@@ -11,6 +11,11 @@ use types::{RawTableCollection, Table, TableCollection};
 
 use crate::parser::parse;
 
+#[cfg(feature = "generate")]
+pub mod generate;
+#[cfg(not(feature = "generate"))]
+mod generate;
+
 mod parser;
 pub mod types;
